+++ src/snowflake/snowpark/mock/_constants.py
@@ -1,0 +1,12 @@
+#
+# Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
+#
+
+# overwrite the following values to change the default information in a mock session
+# if the corresponding fields are not set in the session creation options
+CURRENT_ACCOUNT = "mock_account"
+CURRENT_USER = "mock_user"
+CURRENT_DATABASE = "mock_database"
+CURRENT_SCHEMA = "mock_schema"
+CURRENT_WAREHOUSE = "mock_warehouse"
+CURRENT_ROLE = "mock_role"